refactor(tools)!: split tools.py into a tiered package; add MCP_TIERS + put_json
Reorganize the ~2.4k-line tools.py into a tools/ package with one module per
tier (compute, discovery, data_ops, reports, jobs, automl, model_scoring,
decisioning), each exposing register(mcp, get_token) and depending only on the
shared lower layers plus tools/_common (the session-helper factory) -- no tier
imports another, so any subset can be registered on its own.

- Add MCP_TIERS env var + register_tools(tiers=) to expose only selected tiers
  (e.g. MCP_TIERS=0-4 or 0,1,7); unset means all tiers.
- Extract put_json (GET-for-ETag + If-Match PUT) into viya_client alongside
  get_json/post_json/delete_resource; the three update_* decisioning tools now
  call it instead of repeating the ETag-then-PUT block.
- Move get_mas_module_step_signature into Tier 6 (Model Management & Scoring).

Tool behavior and the register_tools(mcp, get_token) entry point are unchanged.
Tests: conftest/payload patch targets updated for the new module layout; adds
tests/test_tier_selection.py and put_json unit tests. ruff/pyright clean,
260 passed, coverage 94.71%.

BREAKING CHANGE: list_models_and_decisions renamed to list_mas_modules. The
tool lists MAS modules (/microanalyticScore/modules); the new name matches what
it returns and its sibling get_mas_module_step_signature.
