+++ .env.sample
@@ -8,6 +8,12 @@ MCP_SIGNING_KEY=default # Should be a strong random string in production
 MCP_BASE_URL= # External URL of the MCP server (e.g. https://sas-mcp.company.com) — defaults to http://localhost:HOST_PORT
 COMPUTE_CONTEXT_NAME=SAS Job Execution compute context
 
+# Expose only a subset of tool tiers (default: all). Accepts ranges and comma
+# lists, e.g. MCP_TIERS=0-4 or MCP_TIERS=0,1,6,7. Tiers: 0 Compute & Code
+# Execution, 1 Data Discovery, 2 Data Operations & Files, 3 Reports, 4 Batch
+# Jobs, 5 Automated ML, 6 Model Management & Scoring, 7 Decisioning.
+MCP_TIERS=
+
 # Set to false to disable SSL verification (e.g. for self-signed Viya certificates)
 SSL_VERIFY=true
