+++ src/sas_mcp_server/config.py
@@ -117,6 +117,9 @@ CLIENT_ID = os.getenv("CLIENT_ID", "sas-mcp")
 HOST_PORT = int(os.getenv("HOST_PORT", "8134"))
 MCP_SIGNING_KEY = os.getenv("MCP_SIGNING_KEY", "default")
 CONTEXT_NAME = os.getenv("COMPUTE_CONTEXT_NAME", "SAS Job Execution compute context")
+# Optional tool-tier selection, e.g. "0-4" or "0,1,7". Empty means all tiers.
+# Parsed by sas_mcp_server.tools.resolve_enabled_tiers.
+MCP_TIERS = os.getenv("MCP_TIERS", "")
 MCP_BASE_URL = os.getenv("MCP_BASE_URL", f"http://localhost:{HOST_PORT}")
 # Upper bound on binary export bytes returned inline as an embedded resource by
 # ``export_report``. Larger exports are refused with guidance rather than
